requirements.txt
flask==2.2.5
flask_sqlalchemy==3.0.3
flask_migrate==4.0.4
flask_login==0.6.2
python-dotenv==1.0.0
pandas==2.2.2
email-validator==1.3.1
